Python/SOCAR_Hackhathon/HackathonOCR_1/requirements.txt
# =================================================================
# Core Python Utilities
# =================================================================
python-dotenv             # General utility for reading environments (settings.py)
typing_extensions         # Required for TypedDict/Pydantic features on older Python versions
pydantic                  # Data validation and API contracts (api/models.py)
requests                  # General HTTP requests (for API calls, though not strictly required if only using the Azure SDK)

# =================================================================
# 1. OCR and Document Processing
# =================================================================
pymupdf                   # PDF Handling and Image Conversion (pdf_to_image.py)
opencv-python             # Image Preprocessing (Noise Reduction, Binarization) (preprocessors.py)
numpy                     # Required by OpenCV for handling image arrays

# =================================================================
# 2. Azure LLM and Embedding Services
# =================================================================
openai                    # Azure OpenAI SDK for LLM (DeepSeek/GPT-4o) and Embedding (text-embedding-3-large) calls

# =================================================================
# 3. Vector Database and Chunking
# =================================================================
qdrant-client             # Vector Database Client (qdrant_client.py)
# For optimal RAG performance, a dedicated text-splitting library is crucial. 
# Since we designed for Markdown splitting, we include a common RAG utility:
langchain-text-splitters  # Recommended tool for recursive/markdown-aware chunking (chunker_utils.py)

# =================================================================
# 4. API and Web Interface (Hackathon Deliverables)
# =================================================================
fastapi                   # Core web framework (main.py, router.py)
uvicorn                   # ASGI server to run the FastAPI application
python-multipart          # Required by FastAPI to handle file uploads (multipart/form-data)
streamlit                 # Frontend framework (If you choose to use the basic Streamlit interface for the chat agent)